# Drastically speed up imports when bulk_size is specified